+++ src/Calculators/Electronics/Basic/OhmsLaw/OhmsLaw.cs
@@ -59,7 +59,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                         new NumberUnit("V", 1e0, NumberPreference.DEFAULT),
                         new NumberUnit("kV", 1e3),
                     },
-                    2,
+                    4,
                     Directions.Input,
                     null);
 
@@ -92,7 +92,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                         new NumberUnit("mA", 1e-3),
                         new NumberUnit("A", 1e0, NumberPreference.DEFAULT),
                     },
-                    2,
+                    4,
                     Directions.Input,
                     null);
 
@@ -123,7 +123,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                         new NumberUnit("MΩ", 1e6),
                         new NumberUnit("GΩ", 1e9),
                     },
-                    2,
+                    4,
                     Directions.Output,
                     null);