+++ src/Core/CalcVarNumerical.cs
@@ -113,6 +113,10 @@ namespace NinjaCalc {
                         this.ioRadioButton.IsChecked = true;
                     }
 
+                    // Now the event handler is removed, update the displayed value for the raw value
+                    // (this is so rounding occurs)
+                    this.UpdateDispValFromRawVal();
+
                 }
                 else if (value == Directions.Input) {
                     //this.calcValTextBox.IsEnabled = true;
@@ -228,10 +232,6 @@ namespace NinjaCalc {
             // Initialise empty validation results list
             this.ValidationResults = new List<CalcValidationResult>();
 
-            // Setup default direction
-            this.Direction = defaultDirection;
-
-            // Internally save reference to the units combo box
             this.unitsComboBox = unitsComboBox;
 
             // Attach event handler to the selection change for the units combo box
@@ -258,6 +258,14 @@ namespace NinjaCalc {
             else {
                 this.SelUnit = this.units[0];
             }
+
+            // Setup default direction
+            this.Direction = defaultDirection;
+
+            // Internally save reference to the units combo box
+           
+
+            
 
             this.NumDigitsToRound = numDigitsToRound;
 
@@ -302,8 +310,9 @@ namespace NinjaCalc {
             // Invoke the provided equation function,
             // which should return the raw value for this calculator variable
             this.rawVal = this.Equation.Invoke();
-            this.dispVal = this.rawVal / this.selUnit.Multiplier;
-            this.calcValTextBox.Text = this.dispVal.ToString();
+            //this.dispVal = this.rawVal / this.selUnit.Multiplier;
+            //this.calcValTextBox.Text = this.dispVal.ToString();
+            this.UpdateDispValFromRawVal();
 
             // Validation is done in the TextBoxChanged event handler
             this.Validate();
@@ -435,12 +444,17 @@ namespace NinjaCalc {
 
             }
             else if(this.Direction == Directions.Output) {
-                // Recalculate dispVal and update textbox
-                // We don't need to validate again if the units are changed for an input,
-                // as the actual value (raw value) does not change.
-                this.dispVal = this.rawVal / this.selUnit.Multiplier;
-                this.calcValTextBox.Text = this.dispVal.ToString();
+                this.UpdateDispValFromRawVal();
             }
+        }
+
+        private void UpdateDispValFromRawVal() {
+            // Recalculate dispVal and update textbox
+            // We don't need to validate again if the units are changed for an output,
+            // as the actual value (raw value) does not change.
+            double unroundedDispVal = this.rawVal / this.selUnit.Multiplier;
+            this.dispVal = Rounding.RoundToSignificantDigits(unroundedDispVal, this.NumDigitsToRound);
+            this.calcValTextBox.Text = this.dispVal.ToString();
         }
 
         public void UpdateUIBasedOnValidationResults() {