+++ changelog.md
@@ -2,6 +2,7 @@ v0.3.6
 ------
 
 - Added mandatory "category" and "tag" fields to the Calculator class, closes #40.
+- Made the calculator selection grid scrollable, closes #41.
 
 v0.3.5
 ------